REFACTOR:배치 좌표 자료형 Float -> Int로 변경

+++ backend/src/main/java/com/ssafy/dingdong/domain/room/entity/RoomFurniture.java
@@ -23,9 +23,9 @@ public class RoomFurniture {
 	private Long roomFurnitureId;
 	private Long roomId;
 	private String assetId;
-	private Float xPos;
-	private Float yPos;
-	private Float zPos;
+	private int xPos;
+	private int yPos;
+	private int zPos;
 	private Short rotation;
 
 }
